refactor(tts): retire copied OpenAI test provider
